# documents/generated-src/openapi-bundled.yaml
openapi: 3.0.3
info:
  title: Ruid.ai API
  version: 1.0.0
  description: Ruid.ai API一覧
servers:
  - url: http://localhost:3030
tags:
  - name: Question
    description: 問題取得API
  - name: Room
    description: 問題グループ取得API
paths:
  /questions/{question_id}:
    get:
      tags:
        - Question
      summary: 問題取得
      description: |
        指定した問題IDに対応する問題を取得します。

        - 問題文
        - 選択肢
        - 解説
        を返します。

        問題が存在しない場合は404を返します。
      operationId: getQuestion
      parameters:
        - name: question_id
          description: パスパラメータ内のquestion_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          $ref: '#/components/responses/getQuestionResponse'
        '400':
          $ref: '#/components/responses/errorResponse'
        '401':
          $ref: '#/components/responses/errorResponse'
        '404':
          $ref: '#/components/responses/errorResponse'
        '500':
          $ref: '#/components/responses/errorResponse'
  /rooms:
    post:
      tags:
        - Room
      summary: ルーム作成
      description: |
        指定された条件をもとに新しい問題を作成します。

        リクエストボディには、問題ジャンルや問題形式などの生成条件を指定します。

        作成に成功した場合は201 Createdを返し、Locationヘッダーに作成した問題のリソースURIを設定します。
      operationId: postRoom
      requestBody:
        $ref: '#/components/requestBodies/qprompt'
      responses:
        '201':
          $ref: '#/components/responses/createRoomResponse'
        '400':
          $ref: '#/components/responses/errorResponse'
        '401':
          $ref: '#/components/responses/errorResponse'
        '404':
          $ref: '#/components/responses/errorResponse'
        '500':
          $ref: '#/components/responses/errorResponse'
  /rooms/{room_id}:
    get:
      tags:
        - Room
      sumamry: ルーム情報取得
      description: |
        ルームIDから、ルームの情報を取得します。

        - ルーム名
        を返します。

        ルームが存在しない場合は404を返します。
      operationId: getRoom
      parameters:
        - name: room_id
          description: パスパラメータ内のroom_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          $ref: '#/components/responses/getRoomResponse'
        '400':
          $ref: '#/components/responses/errorResponse'
        '401':
          $ref: '#/components/responses/errorResponse'
        '404':
          $ref: '#/components/responses/errorResponse'
        '500':
          $ref: '#/components/responses/errorResponse'
  /rooms/{room_id}/questions:
    get:
      tags:
        - Room
      summary: 問題リスト取得
      description: |
        指定したルームID内の問題リストを取得します。

        - 問題要約
        - 問題ID
        を返します。

        ルームが存在しない場合は404を返します。
      operationId: getQuestionsByRoomId
      parameters:
        - name: room_id
          description: パスパラメータ内のroom_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          $ref: '#/components/responses/getQuestionsByRoomIdResponse'
        '400':
          $ref: '#/components/responses/errorResponse'
        '401':
          $ref: '#/components/responses/errorResponse'
        '404':
          $ref: '#/components/responses/errorResponse'
        '500':
          $ref: '#/components/responses/errorResponse'
  /rooms/{room_id}/questions/next:
    get:
      tags:
        - Room
      sumamry: 次の問題の問題ID取得
      description: |
        ルームIDから、そのルーム内で未回答の問題を取得します。

        - 問題ID
        を返します。

        未回答の問題が存在しない場合は404を返します。
      operationId: getNewQuestionId
      parameters:
        - name: room_id
          description: パスパラメータ内のroom_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          $ref: '#/components/responses/getNewQuestionIdResponse'
        '400':
          $ref: '#/components/responses/errorResponse'
        '401':
          $ref: '#/components/responses/errorResponse'
        '404':
          $ref: '#/components/responses/errorResponse'
        '500':
          $ref: '#/components/responses/errorResponse'
components:
  requestBodies:
    Qprompt:
      $ref: '#/components/requestBodies/qprompt'
    qprompt:
      content:
        application/json:
          schema:
            type: object
            required:
              - prompt
            properties:
              prompt:
                type: string
  responses:
    ErrorResponse:
      $ref: '#/components/responses/errorResponse'
    GetQuestionResponse:
      $ref: '#/components/responses/getQuestionResponse'
    CreateRoomResponse:
      $ref: '#/components/responses/createRoomResponse'
    GetRoomResponse:
      $ref: '#/components/responses/getRoomResponse'
    GetQuestionsByRoomIdResponse:
      $ref: '#/components/responses/getQuestionsByRoomIdResponse'
    GetNewQuestionIdResponse:
      $ref: '#/components/responses/getNewQuestionIdResponse'
    getQuestionResponse:
      description: 問題を取得しました。
      content:
        application/json:
          schema:
            type: object
            required:
              - room_id
              - question_text
              - choices
              - correct_answer_index
              - explanation_text
              - answered_flg
              - summary
            properties:
              room_id:
                type: string
                description: Room ID
                format: uuid
              question_text:
                type: string
                description: 問題文（Markdown）
                example: |
                  テスト用問題文
                  改行もあるよ
                  **how's the bold letter?**
              choices:
                type: array
                description: 選択肢配列. 選択肢を表す記号と選択肢の本文のセット
                items:
                  type: object
                  required:
                    - tag
                    - text
                  properties:
                    tag:
                      type: string
                    text:
                      type: string
                example:
                  - tag: A
                    text: choice1
                  - tag: B
                    text: choice2
              correct_answer_index:
                type: integer
                description: 選択肢の中で、正解のインデックス
                example: 1
              explanation_text:
                type: string
                description: 解説文（Markdown）
                example: |
                  ここに説明が表示されるはずです
                  **それも、マークダウンで！**
              answered_flg:
                type: boolean
                nullable: true
                description: 回答ステータスフラグ
              summary:
                type: string
                description: 問題の短い説明
    errorResponse:
      description: エラー
      content:
        application/json:
          schema:
            type: object
            required:
              - errorCode
              - content
            properties:
              errorCode:
                type: string
                description: エラーコード
                example: VALIDATION_ERROR
              headers:
                type: object
                description: ヘッダー（ここでは使わないが、他の正常レスポンスとの共通化をはかる）
                example: ''
              content:
                type: object
                required:
                  - message
                properties:
                  message:
                    type: string
                    description: エラーメッセージ
                    example: question_idはUUID形式で指定してください。
    createRoomResponse:
      description: 問題を作成しました。
      headers:
        Location:
          description: 生成したルームID
          required: true
          schema:
            type: string
            format: uuid
      content:
        application/json:
          schema:
            type: object
            required:
              - room_id
            properties:
              room_id:
                type: string
                format: uuid
    getRoomResponse:
      description: ルーム情報を取得しました。
      content:
        application/json:
          schema:
            type: object
            required:
              - title
            properties:
              title:
                type: string
                description: ルームの端的な説明＝ルーム名
                example: 応用情報：ストラテジ分野
    getQuestionsByRoomIdResponse:
      description: 問題リストを取得しました。
      content:
        application/json:
          schema:
            type: array
            items:
              type: object
              required:
                - question_id
                - summary
              properties:
                question_id:
                  type: string
                  description: Question ID
                  format: uuid
                summary:
                  type: string
                  description: 問題要約
    getNewQuestionIdResponse:
      description: 新しい問題を取得しました。
      content:
        application/json:
          schema:
            type: object
            required:
              - question_id
            properties:
              question_id:
                type: string
                description: Question ID
                format: uuid
  headers:
    Location:
      $ref: '#/components/headers/location'
    SetCookie:
      $ref: '#/components/headers/set-cookie'
    WWWAuthenticate:
      $ref: '#/components/headers/www-authenticate'
    location:
      description: 作成したリソースのURI
      schema:
        type: string
        format: uri
    set-cookie:
      description: |
        Cookieを設定する。 HttpOnly、Secure、SameSite=Laxなどの属性を含む。
      schema:
        type: string
    www-authenticate:
      description: 認証方式と認証エラー情報
      schema:
        type: string